+++ src/Storage/Storage.Management/ChangeLog.md
@@ -18,6 +18,9 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Supported Zone and ZonePlacementPolicy on Storage accounts 
+    - `New-AzStorageAccount`
+    - `Set-AzStorageAccount`
 * Supported listing Storage SKU
     - `Get-AzStorageSku`
 * Supported enabling SMB Oauth on Storage accounts